deploy/requirements.txt
fastapi
uvicorn
transformers
torch
pytesseract
Pillow
duckduckgo-search
requests
